+++ lib/visual_cache.dart
@@ -1,18 +1,31 @@
 library visual_cache;
 
 import 'dart:io';
+import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 
 class VisualCache {
   /// Gets the cache directory and returns its size in MB or GB.
-  Future<String> getCacheSize() async {
+  /// Gets the specified cache subdirectories and returns their total size in MB or GB.
+  /// If no subdirectories are specified, it returns the size of the entire cache directory.
+  Future<Map<String, int>> getCacheSizeDetails({List<String>? subdirectories}) async {
     final cacheDir = await getTemporaryDirectory();
-    final cacheSize = await _calculateDirectorySize(cacheDir);
+    Map<String, int> sizes = {};
 
-    // Convert size to MB or GB
-    return (cacheSize / (1024 * 1024)) < 1024
-        ? '${(cacheSize / (1024 * 1024)).toStringAsFixed(2)} MB'
-        : '${(cacheSize / (1024 * 1024 * 1024)).toStringAsFixed(2)} GB';
+    if (subdirectories == null || subdirectories.isEmpty) {
+      sizes['Total Cache'] = await _calculateDirectorySize(cacheDir);
+    } else {
+      for (var subdir in subdirectories) {
+        final subDirPath = Directory('${cacheDir.path}/$subdir');
+        if (await subDirPath.exists()) {
+          sizes[subdir] = await _calculateDirectorySize(subDirPath);
+        } else {
+          sizes[subdir] = 0; // Or exclude this line if you don't want to show empty directories
+        }
+      }
+    }
+
+    return sizes;
   }
 
   /// Recursively calculates the size of a directory and its subdirectories.
@@ -26,7 +39,7 @@ class VisualCache {
         }
       }
     } catch (e) {
-      print('Error calculating directory size: $e');
+      debugPrint('Error calculating directory size: $e');
     }
     return size;
   }