+++ CHANGELOG.md
@@ -6,3 +6,14 @@
 
 * Platform support updates
 * Update README.md
+
+## 1.0.0-betta.1
+
+* Now you don’t need to write a grammar construct, now you just need to call the VisualCacheWidget() widget;
+* Added visual cache viewing using the pie_chart 5.4.0 library;
+* Added the ability to add your own directory. Example:    
+``` dart 
+VisualCacheWidget(subdirectories: ['NewFolder','NewFolder2', 'NewFolder3'],)
+```
+* Update README.md;
+* Bug fixes;